Auth-Service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.4",
    "django==5.2.2",
    "django-cors-headers>=4.7.0",
    "django-countries>=7.6.1",
    "django-environ>=0.12.0",
    "django-phonenumber-field>=8.1.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "gunicorn>=23.0.0",
    "phonenumbers>=9.0.6",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "setuptools>=80.9.0",
    "uvicorn[standard]>=0.34.2",
]

